requirements.txt
absl-py==2.1.0
asgiref==3.7.2
autograd==1.6.2
autograd-gamma==0.5.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
DateTime==5.4
Django==5.0.1
fonttools==4.50.0
formulaic==1.0.1
future==1.0.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
idna==3.6
interface-meta==1.3.0
kiwisolver==1.4.5
lifelines==0.28.0
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.8.3
numpy==1.26.3
oauthlib==3.2.2
openslide-python==1.3.1
packaging==24.0
pandas==2.2.0
Pillow==9.5.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.12.0
shutilwhich==1.1.0
six==1.16.0
sqlparse==0.4.4
tensorboard==2.15.1
tensorboard-data-server==0.7.2
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
zope.interface==6.2
